+++ tests/native/test_parallel_client.py
@@ -23,12 +23,11 @@ import unittest
 import pwd
 import os
 import shutil
-import sys
 import string
 import random
 from hashlib import sha256
 from datetime import datetime
-from platform import python_version
+from unittest.mock import patch, MagicMock
 
 from gevent import joinall, spawn, socket, Greenlet, sleep, Timeout as GTimeout
 from pssh.config import HostConfig
@@ -36,7 +35,7 @@ from pssh.clients.native import ParallelSSHClient
 from pssh.exceptions import UnknownHostException, \
     AuthenticationException, ConnectionErrorException, SessionError, \
     HostArgumentException, SFTPError, SFTPIOError, Timeout, SCPError, \
-    PKeyFileError, ShellError, HostArgumentError
+    PKeyFileError, ShellError, HostArgumentError, NoIPv6AddressFoundError
 from pssh.output import HostOutput
 
 from .base_ssh2_case import PKEY_FILENAME, PUB_FILE
@@ -272,13 +271,7 @@ class ParallelSSHClientTest(unittest.TestCase):
         self.assertEqual(len(hosts), len(output))
         self.assertIsNotNone(output[1].exception)
         self.assertEqual(output[1].exception.host, hosts[1])
-        try:
-            raise output[1].exception
-        except ConnectionErrorException:
-            pass
-        else:
-            raise Exception("Expected ConnectionError, got %s instead" % (
-                output[1].exception,))
+        self.assertIsInstance(output[1].exception, ConnectionErrorException)
 
     def test_pssh_client_timeout(self):
         # 1ms timeout
@@ -1894,6 +1887,33 @@ class ParallelSSHClientTest(unittest.TestCase):
             for host_out in output:
                 stdout = list(host_out.stdout)
                 self.assertListEqual(stdout, [self.resp])
+
+    @patch('pssh.clients.base.single.socket')
+    def test_ipv6(self, gsocket):
+        hosts = ['::1']
+        client = ParallelSSHClient(hosts, port=self.port, pkey=self.user_key, num_retries=1)
+        addr_info = ('::1', self.port, 0, 0)
+        gsocket.IPPROTO_TCP = socket.IPPROTO_TCP
+        gsocket.socket = MagicMock()
+        _sock = MagicMock()
+        gsocket.socket.return_value = _sock
+        sock_con = MagicMock()
+        _sock.connect = sock_con
+        getaddrinfo = MagicMock()
+        gsocket.getaddrinfo = getaddrinfo
+        getaddrinfo.return_value = [(
+            socket.AF_INET6, socket.SocketKind.SOCK_STREAM, socket.IPPROTO_TCP, '', addr_info)]
+        output = client.run_command(self.cmd, stop_on_errors=False)
+        for host_out in output:
+            self.assertEqual(hosts[0], host_out.host)
+            self.assertIsInstance(host_out.exception, TypeError)
+
+    def test_no_ipv6(self):
+        client = ParallelSSHClient([self.host], port=self.port, pkey=self.user_key, num_retries=1, ipv6_only=True)
+        output = client.run_command(self.cmd, stop_on_errors=False)
+        for host_out in output:
+            # self.assertEqual(self.host, host_out.host)
+            self.assertIsInstance(host_out.exception, NoIPv6AddressFoundError)
 
     # TODO:
     # * password auth