+++ Changelog.rst
@@ -1,6 +1,17 @@
 Change Log
 ============
 
+2.7.0
++++++
+
+Changes
+-------
+
+* All clients now support IPv6 addresses for both DNS and IP entries in host list.
+* Added ``ipv6_only`` flag to ``ParallelSSHClient`` and ``SSHClient`` for choosing only IPv6 addresses when both v4 and
+  v6 are available.
+* Removed Python 2 from binary wheel compatibility as it is no longer supported and not guaranteed to work.
+
 2.6.0
 +++++